src/cloud-functions/cf1_process_incoming_transactions/requirements.txt
google-cloud-pubsub
google-cloud-firestore==1.9.0